Move some helpers
