+++ src/lib/readable-stream/default-controller.ts
@@ -9,15 +9,10 @@ import {
 import { SimpleQueue } from '../simple-queue';
 import { CancelSteps, PullSteps } from './symbols';
 import { ReadableStreamCreateReadResult, ReadResult } from './generic-reader';
-import {
-  CreateAlgorithmFromUnderlyingMethod,
-  InvokeOrNoop,
-  promiseResolvedWith,
-  typeIsObject,
-  uponPromise
-} from '../helpers';
+import { CreateAlgorithmFromUnderlyingMethod, InvokeOrNoop, promiseResolvedWith, uponPromise } from '../helpers';
 import { IsReadableStreamLocked, ReadableStream, ReadableStreamClose, ReadableStreamError } from '../readable-stream';
 import { UnderlyingSource } from './underlying-source';
+import { typeIsObject } from '../helpers/miscellaneous';
 
 export class ReadableStreamDefaultController<R> {
   /** @internal */
