+++ src/lib/transform-stream.ts
@@ -9,7 +9,6 @@ import {
   promiseRejectedWith,
   promiseResolvedWith,
   transformPromiseWith,
-  typeIsObject,
   ValidateAndNormalizeHighWaterMark
 } from './helpers';
 import { CreateReadableStream, ReadableStream, ReadableStreamDefaultController } from './readable-stream';
@@ -23,6 +22,7 @@ import {
 } from './readable-stream/default-controller';
 import { QueuingStrategy, QueuingStrategySizeCallback } from './queuing-strategy';
 import { CreateWritableStream, WritableStream, WritableStreamDefaultControllerErrorIfNeeded } from './writable-stream';
+import { typeIsObject } from './helpers/miscellaneous';
 
 export type TransformStreamDefaultControllerCallback<O>
   = (controller: TransformStreamDefaultController<O>) => void | PromiseLike<void>;
